# ui/src/api/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "quarry-server",
    "description": "",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.1.7-nightly.20260731"
  },
  "paths": {
...
    "/v1/events": {
      "get": {
        "tags": [
          "sse"
        ],
        "operationId": "events",
        "parameters": [
          {
            "name": "library",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Server-sent event stream"
          }
        }
      }
    },
...
    "/v1/libraries/{library}": {
      "get": {
        "tags": [
          "library_handlers"
        ],
        "operationId": "get_library",
        "parameters": [
          {
            "name": "library",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Library"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiErrorCode": {
        "type": "string",
        "description": "Stable machine-readable codes returned by every `/v1` HTTP error.",
        "enum": [
          "FORBIDDEN",
          "INVALID_REQUEST",
          "NOT_FOUND",
          "GONE",
          "PRECONDITION_REQUIRED",
          "PRECONDITION_FAILED",
          "CONFLICT",
          "METHOD_NOT_ALLOWED",
          "UNSUPPORTED_MEDIA_TYPE",
          "PAYLOAD_TOO_LARGE",
          "UNPROCESSABLE_ENTITY",
          "SERVICE_BUSY",
          "INTERNAL_ERROR",
          "STALE_BASE",
          "BLOCK_DELETED",
          "ANCHOR_NOT_FOUND",
          "BLOCK_MOVE_CONFLICT",
          "SUGGESTION_INVALIDATED",
          "SUGGESTION_ALREADY_RESOLVED",
          "UNSUPPORTED_MARKDOWN",
          "INVALID_TRANSACTION",
          "UNKNOWN_BLOCK_TYPE",
          "UNSUPPORTED_BLOCK_DOCUMENT"
        ]
      },
      "ApiErrorDetails": {
        "type": "object",
        "description": "Machine-actionable context for a failed request. Transaction failures\nidentify the exact op and target; validation failures may additionally\nidentify the field, rejected value, current value, or allowed values.",
        "properties": {
          "allowed_values": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Accepted values for a closed vocabulary."
          },
          "current_value": {
            "type": [
              "string",
              "null"
            ],
            "description": "Current server value, when useful for rebuilding a request."
          },
          "field": {
            "type": [
              "string",
              "null"
            ],
            "description": "Request field that failed validation or a precondition."
          },
          "op": {
            "type": [
              "string",
              "null"
            ],
            "description": "Operation discriminator, such as `replace_block_content`."
          },
          "op_index": {
            "type": [
              "integer",
              "null"
            ],
            "description": "Zero-based index into a transaction's `ops` array.",
            "minimum": 0
          },
          "target": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ApiErrorTarget",
                "description": "Addressable entity involved in the failure."
              }
            ]
          },
          "value": {
            "type": [
              "string",
              "null"
            ],
            "description": "Rejected request value."
          }
        }
      },
      "ApiErrorResponse": {
        "type": "object",
        "required": [
          "code",
          "retryable",
          "message"
        ],
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ApiErrorCode"
          },
          "details": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ApiErrorDetails"
              }
            ]
          },
          "message": {
            "type": "string"
          },
          "retryable": {
            "type": "boolean"
          }
        }
      },
      "ApiErrorTarget": {
        "type": "object",
        "required": [
          "kind",
          "id"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "Stable identifier copied from the request or current review state."
          },
          "kind": {
            "type": "string",
            "description": "Entity category, such as `block`, `review_item`, or `conflict`."
          }
        }
      },
...
      "Library": {
        "type": "object",
        "required": [
          "id",
          "slug",
          "created_at",
          "settings"
        ],
        "properties": {
          "created_at": {
            "$ref": "#/components/schemas/Timestamp"
          },
          "id": {
            "type": "string"
          },
          "settings": {
            "$ref": "#/components/schemas/Value"
          },
          "slug": {
            "type": "string"
          }
        }
      },
...
      "Timestamp": {
        "type": "string"
      },
...
      "Value": {},
...
    }
  }
}